[iOS] feat: 멸균실 네비게이션 타이틀 기능 구현 (#56)
