+++ Mondee/View/CleanRoomView/CleanRoomView.swift
@@ -12,62 +12,24 @@ struct CleanRoomView: View {
     @State private var scrollViewOffset: CGFloat = 0
     
     var body: some View {
-        ZStack {
-            Color.mondeeBoxBackground.ignoresSafeArea()
-            ScrollView {
-                ZStack {
-                    Color.gray.ignoresSafeArea()
-                    VStack {
-                        ZStack {
-                            Header
-                                .opacity(1 - Double(-scrollViewOffset) / 90)
-                                .onScrollViewOffsetChanged { value in
-                                    self.scrollViewOffset = value
-                                }
-                        }
-                        Rectangle()
-                            .frame(height: 400)
-                            .foregroundColor(.white)
-                            .padding(.horizontal)
-                        Rectangle()
-                            .frame(height: 400)
-                            .foregroundColor(.white)
-                            .padding(.horizontal)
-                        Rectangle()
-                            .frame(height: 400)
-                            .foregroundColor(.white)
-                            .padding(.horizontal)
-                    }
-                }
-            }
-            .overlay (Text("\(scrollViewOffset)"))
-            .overlay(alignment: .top) {
-                navigationBar.opacity(scrollViewOffset < -60 ? 1 : 0)
-                    .animation(.easeInOut, value: scrollViewOffset)
-            }
+        ScrollView {
+            LargeNavigationTitle(scrollViewOffset: $scrollViewOffset)
+            
+        }
+        .edgesIgnoringSafeArea(.top)
+//        .overlay (Text("\(scrollViewOffset)"))
+        .overlay(alignment: .top) {
+            InlineNavigationTitle(scrollViewOffset: scrollViewOffset)
         }
     }
 }
 
-extension CleanRoomView {
+struct LargeNavigationTitle: View {
     
-    private var navigationBar: some View {
-        VStack(alignment: .leading) {
-            Text("멸균실")
-                .font(.headline)
-                .foregroundColor(.mondeeBlue)
-        }
-        .frame(maxWidth: .infinity, alignment: .center)
-        .padding(.bottom)
-        .background(
-            RoundedRectangle(cornerRadius: 22, style: .continuous)
-                .fill(Color.mondeeBoxBackground)
-                .ignoresSafeArea()
-        )
-    }
+    @Binding var scrollViewOffset: CGFloat
     
-    private var Header: some View {
-        VStack(alignment: .leading) {
+    var body: some View {
+        VStack(alignment: .leading, spacing: 5) {
             Text("멸균실")
                 .font(.title)
                 .fontWeight(.bold)
@@ -75,48 +37,44 @@ extension CleanRoomView {
             Text("획득한 먼지들을 확인해보세요")
         }
         .frame(maxWidth: .infinity, alignment: .leading)
-        .padding(.horizontal, 30)
-        .background(
-            Rectangle()
-                .fill(Color.mondeeBoxBackground)
-                .frame(maxWidth: .infinity)
-        )
-        .padding(.bottom, 30)
+        .opacity(1 - Double(-scrollViewOffset) / 60)
+        .padding(EdgeInsets(top: 75, leading: 40, bottom: 20, trailing: 40))
         .background(
             RoundedRectangle(cornerRadius: 22, style: .continuous)
                 .fill(Color.mondeeBoxBackground)
                 .ignoresSafeArea()
         )
+        .onScrollViewOffsetChanged { value in
+            scrollViewOffset = value
+        }
+        .opacity(scrollViewOffset < -55 ? 0 : 1)
     }
 }
 
-struct ScrollViewOffsetPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
+struct InlineNavigationTitle: View {
     
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-extension View {
+    var scrollViewOffset: CGFloat
     
-    func onScrollViewOffsetChanged(action: @escaping (_ offset: CGFloat) -> Void) -> some View {
-        self
-            .background(
-                GeometryReader {geo in
-                    Text("")
-                        .preference(key: ScrollViewOffsetPreferenceKey.self, value: geo.frame(in: .global).minY)
-                }
-            )
-            .onPreferenceChange(ScrollViewOffsetPreferenceKey.self) { value in
-                action(value)
-            }
-        
+    var body: some View {
+        VStack(alignment: .leading) {
+            Text("멸균실")
+                .font(.title3)
+                .fontWeight(.bold)
+                .foregroundColor(.mondeeBlue)
+        }
+        .frame(maxWidth: .infinity, alignment: .center)
+        .padding(.vertical, 15)
+        .background(
+            RoundedRectangle(cornerRadius: 22, style: .continuous)
+                .fill(Color.mondeeBoxBackground)
+                .ignoresSafeArea()
+        )
+        .opacity(scrollViewOffset < -55 ? 1 : 0)
     }
 }
 
 struct CleanRoomView_Previews: PreviewProvider {
     static var previews: some View {
-        CleanRoomView()
+        CleanRoomView().background(Color.mondeeBackgroundGrey)
     }
 }
