+++ Mondee/ViewModel/CleanViewModel.swift
@@ -1,0 +1,31 @@
+//
+//  CleanViewModel.swift
+//  Mondee
+//
+//  Created by Hyunjun Kim on 2023/07/25.
+//
+
+import SwiftUI
+
+struct ScrollViewOffsetPreferenceKey: PreferenceKey {
+    static var defaultValue: CGFloat = 0
+    
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value = nextValue()
+    }
+}
+
+extension View {
+    func onScrollViewOffsetChanged(action: @escaping (_ offset: CGFloat) -> Void) -> some View {
+        self
+            .background(
+                GeometryReader {geo in
+                    Text("")
+                        .preference(key: ScrollViewOffsetPreferenceKey.self, value: geo.frame(in: .global).minY)
+                }
+            )
+            .onPreferenceChange(ScrollViewOffsetPreferenceKey.self) { value in
+                action(value)
+            }
+    }
+}
